data/server/requirements.txt
flask
requests
flask-accept
pyyaml
